Fix enumeration of multiboot info
Previously, the enumeration code used the size provided in the header,
but didn't take into account the fact that this included the size of the
header itself, which caused undefined behavior when reading past the end
of the end of the info. Now, the code simply identifies the terminating
tag (type 0) and stops then.